Time: 1 ms (85.66%), Space: 41 MB (21.72%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int mySqrt(int x) {
+    public int mySqrt(long x) {
         
         long left = 1;
         long right = x + 1;
